+++ src/common/hooks/index.js
@@ -1,1 +1,2 @@
+export * from './useIntegrationConfigs';
 export * from './useTranslatedCategories';